+++ src/assets/env.template.js
@@ -13,4 +13,5 @@
     window['env']['CACHET_URL'] = '${CACHET_URL}'
     window['env']['CACHET_WATCHED_COMPONENTS'] = '${CACHET_WATCHED_COMPONENTS}'
     window['env']['SCHEDULE_LOOKAHEAD_DAYS'] = '${SCHEDULE_LOOKAHEAD_DAYS}'
+    window['env']['FALLBACK_SCHEDULES_URL'] = '${FALLBACK_SCHEDULES_URL}'
 })(this)